src/requirements.txt
gitpython==3.1.30
PyDriller==2.5
python-dotenv==1.0.0
dockerfile==3.3.1
parameter-expansion-patched==0.3.1
pygithub==1.59.0
numpy==1.25.1
matplotlib==3.7.2
ghlinguist==1.1.0
pyyaml==5.4.1
